fixed 2 compiler warnings

+++ AncientArcher/src/renderer/Renderer.h
@@ -18,7 +18,7 @@ public:
 
 private:
 
-  int width;
+  float width;
 
 };
